requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.0
cryptography==2.8
flask==1.1.1
gevent-eventemitter==2.1
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
multidict==4.7.4
protobuf==3.11.2
pycparser==2.19
pyee==6.0.0
requests==2.22.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
steam==0.9.1
urllib3==1.25.7
vdf==3.2
werkzeug==0.16.0
yarl==1.4.2